+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * [THORN-437] Deprecated external bandpass loadbalancer.
   * Added a deprecated compatibility flag `bandpassLoadBalancer.enabled` for bandpass DAQ LoadBalancer Services. Default is `true` to preserve existing behaviour.
+  * Added attribute `bandpassLoadBalancerEnabled` to indicate to indicate to other Tango devices which network to use.
   * New deployments should set it to `false` unless external LoadBalancer exposure is required, for example:
 
         helm_values: